Simplify file installation.
First, check to see if the file has changed.  If it hasn't, don't mess
with the file so that timestamps on the file stay the same.
Otherwise, write content to temporary file in same directory as final
location, so that any selinux labels will be right.  Then move the
file to it's final destination.