Migrates over to using SQLAlchemy-Migrate
Instructions for use are in the README.  This patch does the change so that version 0 of the database assumes
we are using the DB version 0.8.5.  Any upgrade scripts will assume that said fas2.sql has already been deployed
and that all changes operate on top of that.

If we make our changes more DB generic, then we can replace fas2.sql with something also more DB generic.  This will give us the option of migrating away from Postgres if we ever have to.