Reverting previous patch in list and adding a dump method
The new dump method will allow us to optimize changes for fasClient without having to worry about the implications on the rest of the application.  In this case we've moved some of the logic fasClient used to do and are now doing it in the web app.  Initial tests show that the json produced went from 11M to a max of 476K for fedorapeople.  The most common, like that of app1, is around 21K.  This is significantly faster (many times / second vs 1 time / minute).