Make supported languages a config parameter set
Add available_languages, a list to fas.cfg.  This is a replacement for the
hardwired values in ./fas/util.py: available_languages().  This reverts to
previous behavior if available_languages is not defined in fas.cfg.

Add default_language, a string, to fas.cfg.  Add code to ./fas/controllers.py
to use this in get_locale.

fas/templates/master.html will only show a language selection if there are
more than one language defined.  This is in prep for local customization by
way of a single available language for UI text.

Signed-off-by: Jim Lieb <lieb@sea-troll.net>
Signed-off-by: Ricky Zhou<ricky@fedoraproject.org>