Rework openID
From this change we allow OpenID client to hit against
accounts/openid/yadis/ instead of /accounts/openid/yadis/<username>
and still be able to authenticate.

The first change allows to call accounts/openid/yadis/ directly
without running into an error 500 and allowing discovery on the client
side.

The second change just allows to authenticate for a username even if
the username was not provided at first.