Adjust fasClient to load the configuration raw
For gitolite3, we need to specify the username in the command inserted
in ~/.ssh/authorized_keys.
For this we have currently the configuration keys: ssh_restricted_app
and ssh_admin_app but we cannot set them as template using %(username)s
in it.

With this change we now can. As we fill in the command using a dictionary
python will no complain if the string does not contain %(username)s
so it should be backward compatible.

The change to read the configuration file with RawConfigParser is also
safe as afaik we do not interpolate anywhere variables based on others.