Work around Python's invalid ISO format
See the comment above `fix_utc_iso_format` in `fas/utils/__init__.py`,
but the general idea is that Python's datetime carries no timezone
information. So when it's used for UTC, it breaks the
specification (which says that it, if no `Z` or `+HHMM` or `-HHMM` is
specified, then the time should be treated as local time).

Furthermore, since it's invalid, it makes it impossible to correctly
parse in environments which are more strict about following the ISO
format, since such environments are timezone-aware and won't know which
timezone to tag the timestamp with.

Note that this fix is very unideal: Since Python is unityped, there's no
way to ensure that the object passed to `fix_utc_iso_format` is really a
UTC time. This means that if a local time is passed to
`fix_utc_iso_format`, we will once again break the ISO spec and render
it as a UTC timestamp. (If it's not a `datetime` object at all, who
knows what runtime error will happen, but this issue is everywhere in
Python, not specific to `fix_utc_iso_format`.)

Signed-off-by: Ricky Elrod <ricky@elrod.me>