Add fedmenu!
This will add a little menu on the bottom right hand corner of the
screen to link the user to other apps in Fedora Infrastructure.
If the fedmenu.url value is not defined in the config, then nothing is
display (so FAS can be reused other places that don't want to use this
shim).

As a nice bonus, on user profile pages we'll also display a little
button with their avatar than takes you to views of that user in other
infra apps (so you can quickly jump to see what packages they own or
what their datagrepper activity is, etc..).