Store and show only GPG fingerprint instead of ID
The GPG ID should not be trusted and is just a certain amount of bytes
from the fingerprint. Therefore is is redundant to store both the
fingerprint and ID. However, GPG also supports using the full
fingerprint instead of the ID, therefore it is more secure to always use
the GPG fingerprint. See also: https://evil32.com/