Ensure the number of pages returned is always at least 1
Otherwise we may end up in a situation where the number of pages is 0
while the page requested is 1 (the default value).
This allows to show 'There is no groups/no users' instead of returning
a 400 error.

Amend groups.paging since we no longer need the `or 1` since
compute_list_pages_from is fixed