Add two IRC-related fields to the people model
Fedora Hubs wants to offer an easy IRC experience for users who opt in
to it. This adds two fields to the ``People`` model. ``irc_enabled`` is
the flag to indicate whether the user wants to use the Fedora-provided
IRC bouncer, and ``irc_password`` is the Freenode password used to
identify the user. ``ircnick`` will be used to determine the nick to use
and ``email`` can be used for the registration email.

For more context, see:

* https://pagure.io/fedora-hubs/issue/2
* https://pagure.io/fedora-hubs/issue/290
* https://pagure.io/fedora-hubs/issue/291

Signed-off-by: Jeremy Cline <jeremy@jcline.org>